refactor: own CLI model display in runtime
